requirements.txt
google-api-python-client==2.110.0
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
